Update README.md
The code has been changed due to an error in the original version of the code. The changes do not affect the main results or conclusions. Specifically, in the updated code, we decrease the PV population counts by 1 in lines 835 and 866. In order to compensate for resultant changes in the PV excitability, PV gap junction conductance (gelec) is increased to from 0.01 to 0.05 (line 760), PV applied current is increased from 0 to 0.1 (lines 837 and 868) and GABAa conductance between PV groups (gfsX) in increased from 0.25 to 0.5.